+++ bench.txt
@@ -1,3 +1,2 @@
-Array.prototype.splice + Array.prototype[Symbol.iterator] x 5,395,276 ops/sec ±2.17% (88 runs sampled)
-Seeq.prototype[Symbol.iterator] x 4,271,417 ops/sec ±1.47% (89 runs sampled)
-SeeqLinked.prototype[Symbol.iterator] x 9,493,836 ops/sec ±1.13% (90 runs sampled)
+Array.prototype.splice + Array.prototype[Symbol.iterator] x 5,020,434 ops/sec ±1.20% (89 runs sampled)
+Seeq.prototype[Symbol.iterator] x 8,642,069 ops/sec ±1.09% (88 runs sampled)
