+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "standard --fix | snazzy",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "npm run lint && tap",
     "bench": "node bench.js > bench.txt && cat bench.txt"
   },
   "repository": {
