+++ test.js
@@ -1,31 +1,67 @@
 'use strict'
 
-// const Seeq = require('./')
-const Seeq = require('./linked')
+const { test } = require('tap')
+const Seeq = require('./')
+const symbols = require('./symbols')
 
-const size = 3
-const s = new Seeq(size)
+test('constructor', t => {
+  const size = 3
+  const s = new Seeq(size)
+  t.is(s.size, size)
+  t.is(s[symbols.kHead].id, 0)
+  let node = s[symbols.kHead].next
+  for (let i = 1; i < size; i++) {
+    t.is(node.id, i)
+    node = node.next
+  }
+  t.end()
+})
 
-let iteration = 1
+test('push - pop', t => {
+  const size = 3
+  const s = new Seeq(size)
+  s.push(1)
+  t.is(1, s.pop())
+  t.is(undefined, s.pop())
+  t.true(s.done)
+  t.end()
+})
 
-for (let i = 0; i < size; i++) {
-  s.push(`${iteration}-${i}`)
-}
+test('overflow', t => {
+  const size = 3
+  const s = new Seeq(size)
+  for (let i = 0; i < size; i++) {
+    s.push(i)
+  }
+  t.true(s[symbols.kOverflowed])
+  t.false(s.done)
+  t.is(s.pop(), s[symbols.kHead].value)
+  t.false(s[symbols.kOverflowed])
+  t.false(s.done)
+  t.end()
+})
 
-console.log(s.pop())
+test('iterator', t => {
+  const size = 10
+  const s = new Seeq(size)
 
-for (const item of s) {
-  console.log(item)
-}
+  for (let i = 0; i < size; i++) {
+    s.push(i)
+  }
 
-iteration++
+  let count = 0
+  for (const item of s) {
+    t.is(item, count++)
+  }
 
-for (let i = 0; i < size; i++) {
-  s.push(`${iteration}-${i}`)
-}
+  for (let i = 1; i < size - 2; i++) {
+    s.push(count + i)
+  }
 
-console.log(s.pop())
-
-for (const item of s) {
-  console.log(item)
-}
+  const start = size + 1
+  count = 0
+  for (const item of s) {
+    t.is(item, start + count++)
+  }
+  t.end()
+})
